third_party/WebKit/LayoutTests/fast/scrolling/fractional-scroll-offset-document-expected.txt

Verifies that document scrolling supports fractional offset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS document.scrollingElement.scrollTop is within 0.00001 of 0
PASS document.scrollingElement.scrollLeft is within 0.00001 of 0
PASS document.scrollingElement.scrollTop is within 0.00001 of 10
PASS document.scrollingElement.scrollLeft is within 0.00001 of 20
PASS document.scrollingElement.scrollTop is within 0.00001 of 31
PASS document.scrollingElement.scrollLeft is within 0.00001 of 9
PASS document.scrollingElement.scrollTop is within 0.00001 of 0
PASS document.scrollingElement.scrollLeft is within 0.00001 of 0
